+++ docs/changelog.rst
@@ -9,6 +9,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 Unreleased
 ----------
 
+[0.12.2] 2025-01-31
+-------------------
+
 Changed
 ~~~~~~~